131 - place order

+++ src/main/java/vn/hoidanit/laptopshop/domain/OrderDetail.java
@@ -14,7 +14,7 @@ public class OrderDetail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
-    private long quantity;
+    private int quantity;
     private double price;
 
     @ManyToOne
@@ -33,11 +33,11 @@ public class OrderDetail {
         this.id = id;
     }
 
-    public long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(long quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
